+++ src/postulo/notifications/email.py
@@ -12,7 +12,7 @@ from django.core.mail import send_mail
 from django.utils.translation import gettext as _
 from django.utils.translation import gettext_lazy as _lazy
 
-from postulo.plugins.base import FieldSpec, TestResult, declares, shipped
+from postulo.plugins.api import FieldSpec, TestResult, declares, shipped
 
 from .base import Notification